leap: support for SAF SaTimeT timeout variable in MDS transport [#1658] Issue : In current implementation their is a mismatch of variable sizes between SAF API accepting SaTimeT variable type for timeout ( 64 bit) and MDS transport implementation accepting (32 bit) timeout variable , the MDS transport should adopt the size of the timeout parameter from 32 bits to 64 bits.
Fix :
Now MDS provide interface to accept SaTimeT type timeout variable from SAF APIS

+++ osaf/libs/core/common/include/osaf_poll.h
@@ -43,7 +43,7 @@ extern "C" {
  *
  * The return value will always be in the range [0, i_nfds].
  */
-extern unsigned osaf_poll(struct pollfd* io_fds, nfds_t i_nfds, int i_timeout);
+extern unsigned osaf_poll(struct pollfd* io_fds, nfds_t i_nfds, int64_t i_timeout);
 
 /**
  * @brief Wait for events on file descriptors
@@ -88,7 +88,7 @@ extern unsigned osaf_ppoll(struct pollfd* io_fds, nfds_t i_nfds,
  *
  * The return value from this function will always be in the range [-1, 1].
  */
-extern int osaf_poll_one_fd(int i_fd, int i_timeout);
+extern int osaf_poll_one_fd(int i_fd, int64_t  i_timeout);
 
 #ifdef  __cplusplus
 }
